+++ packages/runtime/fixtures/composerApp/platformatic.composer.json
@@ -1,0 +1,23 @@
+{
+  "$schema": "https://platformatic.dev/schemas/v0.23.2/composer",
+  "server": {
+    "hostname": "127.0.0.1",
+    "port": 0,
+    "logger": {
+      "level": "info"
+    }
+  },
+  "composer": {
+    "services": [
+      {
+        "id": "api1",
+        "origin": "http://127.0.0.1",
+        "openapi": {
+          "file": "./api1.json"
+        }
+      }
+    ],
+    "refreshTimeout": 1000
+  },
+  "watch": false
+}
