+++ packages/runtime/fixtures/dbApp/platformatic.db.json
@@ -1,0 +1,22 @@
+{
+  "$schema": "https://platformatic.dev/schemas/v0.22.0/db",
+  "server": {
+    "hostname": "127.0.0.1",
+    "port": 0
+  },
+  "migrations": {
+    "dir": "migrations",
+    "table": "versions"
+  },
+  "types": {
+    "autogenerate": false
+  },
+  "db": {
+    "connectionString": "sqlite://db.sqlite",
+    "graphql": true,
+    "ignore": {
+      "versions": true
+    },
+    "events": false
+  }
+}
