+++ packages/start/package.json
@@ -18,11 +18,7 @@
   },
   "homepage": "https://github.com/platformatic/platformatic#readme",
   "dependencies": {
-    "@platformatic/config": "workspace:*",
-    "@platformatic/db": "workspace:*",
-    "@platformatic/service": "workspace:*",
-    "@platformatic/composer": "workspace:*",
-    "minimist": "^1.2.8"
+    "@platformatic/runtime": "workspace:*"
   },
   "devDependencies": {
     "c8": "^7.13.0",
